% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupyter_utils.R
\name{jupy_table}
\alias{jupy_table}
\title{get a pretty table that can be displayed in Jupyter notebooks}
\usage{
jupy_table(df, digits = 3)
}
\value{
nothing
examples
jupy_table(mtcars,digits=2)
}
\description{
get a pretty table that can be displayed in Jupyter notebooks
}
